+++ content/themes_genesis/atmosphere-pro/page_landing.php
@@ -16,7 +16,19 @@ add_filter( 'body_class', 'atmosphere_add_body_class' );
 function atmosphere_add_body_class( $classes ) {
 
 	$classes[] = 'atmosphere-landing';
+
 	return $classes;
+
+}
+
+//* Remove Skip Links from a template
+remove_action ( 'genesis_before_header', 'genesis_skip_links', 5 );
+
+//* Dequeue Skip Links Script
+add_action( 'wp_enqueue_scripts', 'atmosphere_dequeue_skip_links' );
+function atmosphere_dequeue_skip_links() {
+
+	wp_dequeue_script( 'skip-links' );
 
 }