+++ content/themes_genesis/digital-pro/lib/customize.php
@@ -43,6 +43,7 @@ function digital_customizer_register() {
 
 	$wp_customize->add_setting( 'digital-front-image', array(
 		'default'  => sprintf( '%s/images/front-page-1.jpg', get_stylesheet_directory_uri() ),
+		'sanitize_callback' => 'esc_url_raw',
 		'type'     => 'option',
 	) );