+++ content/themes_genesis/digital-pro/lib/theme-defaults.php
@@ -21,7 +21,7 @@ function digital_theme_defaults( $defaults ) {
 add_action( 'after_switch_theme', 'digital_theme_setting_defaults' );
 function digital_theme_setting_defaults() {
 
-	if( function_exists( 'genesis_update_settings' ) ) {
+	if ( function_exists( 'genesis_update_settings' ) ) {
 
 		genesis_update_settings( array(
 			'blog_cat_num'              => 3,