+++ content/themes_genesis/jessica/lib/plugins/plugins.php
@@ -6,14 +6,15 @@
  * It is expected that theme authors would copy and paste this code into their
  * functions.php file, and amend to suit.
  *
+ * @see http://tgmpluginactivation.com/configuration/ for detailed documentation.
+ *
  * @package    TGM-Plugin-Activation
  * @subpackage Example
- * @version    2.4.0
- * @author     Thomas Griffin <thomasgriffinmedia.com>
- * @author     Gary Jones <gamajo.com>
- * @copyright  Copyright (c) 2014, Thomas Griffin
+ * @version    2.5.2
+ * @author     Thomas Griffin, Gary Jones, Juliette Reinders Folmer
+ * @copyright  Copyright (c) 2011, Thomas Griffin
  * @license    http://opensource.org/licenses/gpl-2.0.php GPL v2 or later
- * @link       https://github.com/thomasgriffin/TGM-Plugin-Activation
+ * @link       https://github.com/TGMPA/TGM-Plugin-Activation
  */
 
 /**